Criando entidade de transacao, rotas e service

+++ backend/src/main/java/io/github/devopMarkz/backend/financas/application/dto/transacao/TransacaoRequestDTO.java
@@ -3,6 +3,7 @@ package io.github.devopMarkz.backend.financas.application.dto.transacao;
 import io.github.devopMarkz.backend.financas.domain.model.Tipo;
 import jakarta.validation.constraints.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class TransacaoRequestDTO {
@@ -15,7 +16,7 @@ public class TransacaoRequestDTO {
     private String descricao;
 
     @PositiveOrZero(message = "Valor deve ser positivo.")
-    private Double valor;
+    private BigDecimal valor;
 
     private LocalDate dataTransacao;
 
@@ -28,7 +29,7 @@ public class TransacaoRequestDTO {
     public TransacaoRequestDTO() {
     }
 
-    public TransacaoRequestDTO(Long categoriaId, String descricao, Double valor, LocalDate dataTransacao, Tipo tipo) {
+    public TransacaoRequestDTO(Long categoriaId, String descricao, BigDecimal valor, LocalDate dataTransacao, Tipo tipo) {
         this.categoriaId = categoriaId;
         this.descricao = descricao;
         this.valor = valor;
@@ -36,7 +37,7 @@ public class TransacaoRequestDTO {
         this.tipo = tipo;
     }
 
-    public TransacaoRequestDTO(Long categoriaId, String descricao, Double valor, LocalDate dataTransacao, Tipo tipo, String observacoes) {
+    public TransacaoRequestDTO(Long categoriaId, String descricao, BigDecimal valor, LocalDate dataTransacao, Tipo tipo, String observacoes) {
         this.categoriaId = categoriaId;
         this.descricao = descricao;
         this.valor = valor;
@@ -45,7 +46,7 @@ public class TransacaoRequestDTO {
         this.observacoes = observacoes;
     }
 
-    public TransacaoRequestDTO(Long categoriaId, String descricao, Double valor, LocalDate dataTransacao, Tipo tipo, Boolean paga) {
+    public TransacaoRequestDTO(Long categoriaId, String descricao, BigDecimal valor, LocalDate dataTransacao, Tipo tipo, Boolean paga) {
         this.categoriaId = categoriaId;
         this.descricao = descricao;
         this.valor = valor;
@@ -54,7 +55,7 @@ public class TransacaoRequestDTO {
         this.paga = paga;
     }
 
-    public TransacaoRequestDTO(Long categoriaId, String descricao, Double valor, LocalDate dataTransacao, Tipo tipo, String observacoes, Boolean paga) {
+    public TransacaoRequestDTO(Long categoriaId, String descricao, BigDecimal valor, LocalDate dataTransacao, Tipo tipo, String observacoes, Boolean paga) {
         this.categoriaId = categoriaId;
         this.descricao = descricao;
         this.valor = valor;
@@ -80,11 +81,11 @@ public class TransacaoRequestDTO {
         this.descricao = descricao;
     }
 
-    public Double getValor() {
+    public BigDecimal getValor() {
         return valor;
     }
 
-    public void setValor(Double valor) {
+    public void setValor(BigDecimal valor) {
         this.valor = valor;
     }
